removed kubelet config features
